requirements.txt
pydantic>=2.6,<3
PyYAML>=6.0.1,<7
